put minian_preview() function into minian_main.py so it can use the new functions and remove minian_preview.py
